fix: separate inference and web for faster web cold start
